requirements.txt
opencv-python
numpy
scikit-image
pywavelets
scikit-learn
joblib
Pillow
matplotlib